+++ scripts/legacyMacros.js
@@ -431,6 +431,7 @@ export {dispatersSupremacy} from './macros/2014/classFeatures/illrigger/painkill
 export {youDieOnMyCommand} from './macros/2014/classFeatures/illrigger/painkiller/youDieOnMyCommand.js';
 export {deathstrike} from './macros/2014/classFeatures/illrigger/painkiller/deathstrike.js';
 export {masterOfTactics} from './macros/2014/classFeatures/rogue/mastermind/masterOfTactics.js';
+export {insightfulManipulator} from './macros/2014/classFeatures/rogue/mastermind/insightfulManipulator.js';
 // Race Features
 export {aasimarNecroticShroud, aasimarRadiantConsumption, aasimarRadiantSoul, celestialRevelation} from './macros/2014/raceFeatures/aasimar/celestialRevelation.js';
 export {acidBreathWeapon, coldBreathWeapon, fireBreathWeapon, lightningBreathWeapon, poisonBreathWeapon} from './macros/2014/raceFeatures/dragonborn/chromatic/breathWeapon.js';