+++ scripts/macros/2024/monsterFeatures/goblinWarrior/scimitar.js
@@ -1,0 +1,9 @@
+export let goblinWarriorScimitar = {
+    name: 'Scimitar',
+    version: '1.3.130',
+    rules: 'modern',
+    monsters: [
+        'Goblin Warrior',
+        'Goblin Boss'
+    ]
+};