+++ scripts/macros/2024/monsterFeatures/worg/bite.js
@@ -1,0 +1,8 @@
+export let worgBite = {
+    name: 'Bite',
+    version: '1.3.130',
+    rules: 'modern',
+    monsters: [
+        'Worg'
+    ]
+};